Refactored code to cleanup 'ContainerView'.
