+++ web-src/src/components/FavoritesSection.js
@@ -12,22 +12,63 @@
 
 import React, { useState } from 'react';
 import { View, IllustratedMessage, Heading, Content, Flex, Tooltip, TooltipTrigger, ActionButton, Well, CheckboxGroup, Checkbox, ActionBarContainer, ActionBar, Item, Text } from '@adobe/react-spectrum';
+import { ToastQueue } from '@react-spectrum/toast'
+
 import Copy from '@spectrum-icons/workflow/Copy';
 import Delete from '@spectrum-icons/workflow/Delete';
 import Edit from '@spectrum-icons/workflow/Edit';
 
 import WriteIcon from '../icons/WriteIcon';
+import { LOCAL_STORAGE_KEY } from '../constants/Constants';
 
-function FavoritesSection({favorites, onCopy, onDelete, onBulkCopy, onBulkDelete}) {
+function FavoritesSection({favorites, onChange}) {
   const [selectedKeys, setSelectedKeys] = useState([]);
 
+  // Function to copy a favorite to the clipboard
+  const copyFavorite = (id) => {
+    const favorite = favorites.find(favorite => favorite.id === id);
+    navigator.clipboard.writeText(favorite.content);
+
+    ToastQueue.positive('Variation copied to clipboard!', {timeout: 2000})
+  }
+
+  // Function to copy all the selected favorites to the clipboard
+  const copySelectedFavorites = (keys) => {
+    const selectedFavorites = favorites.filter(favorite => keys.includes(favorite.id));
+    const selectedFavoritesContent = selectedFavorites.map(favorite => favorite.content);
+    const selectedFavoritesContentString = selectedFavoritesContent.join('\r\n');
+    navigator.clipboard.writeText(selectedFavoritesContentString);
+
+    ToastQueue.positive('Selected saved variations copied successfully!', {timeout: 2000})
+  }
+
+  // Function to delete a favorite and update local storage
+  const deleteFavorite = (id) => {
+    // Filter out the item with the matching id
+    const updatedFavorites = favorites.filter(favorite => favorite.id !== id);
+    onChange(updatedFavorites);
+
+    // Update the local storage with the new favorites
+    localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(updatedFavorites));
+  };
+
+  // Function to delete all the selected favorites and update local storage
+  const deleteSelectedFavorites = (keys) => {
+    const updatedFavorites = favorites.filter(favorite => !keys.includes(favorite.id));
+    onChange(updatedFavorites);
+    localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(updatedFavorites));
+
+    ToastQueue.positive('Selected saved variations deleted successfully!', {timeout: 2000})
+  };
+
+  // Function to handle the action bar actions
   const actionCallback = (key) => {
     if (key === 'select-all') {
       setSelectedKeys(favorites.map((favorite) => favorite.id));
     } else if (key === 'copy-all') {
-      onBulkCopy(selectedKeys);
+      copySelectedFavorites(selectedKeys);
     } else if (key === 'delete-all') {
-      onBulkDelete(selectedKeys);
+      deleteSelectedFavorites(selectedKeys);
     }
   }
   
@@ -72,13 +113,13 @@ function FavoritesSection({favorites, onCopy, onDelete, onBulkCopy, onBulkDelete
                     >
                       <Flex direction="row" gap="size-100" justifyContent="right">
                         <TooltipTrigger delay={0}>
-                          <ActionButton isQuiet UNSAFE_className="hover-cursor-pointer" onPress={() => onCopy(favorite.id)}>
+                          <ActionButton isQuiet UNSAFE_className="hover-cursor-pointer" onPress={() => copyFavorite(favorite.id)}>
                             <Copy />
                           </ActionButton>
                           <Tooltip>Copy to Clipboard</Tooltip>
                         </TooltipTrigger>
                         <TooltipTrigger delay={0}>
-                          <ActionButton isQuiet UNSAFE_className="hover-cursor-pointer" onPress={() => onDelete(favorite.id)}>
+                          <ActionButton isQuiet UNSAFE_className="hover-cursor-pointer" onPress={() => deleteFavorite(favorite.id)}>
                             <Delete />
                           </ActionButton>
                           <Tooltip>Delete Saved Variation</Tooltip>
